app/requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
python-jose